Use trakt endpoint to fetch data & do some cleanup.

+++ shared/domain/episodes/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/episodes/api/ObserveEpisodesInteractor.kt
@@ -22,7 +22,7 @@ class ObserveEpisodesInteractor constructor(
 }
 
 data class EpisodeQuery(
-    val tvShowId: Long,
-    val seasonId: Long,
-    val seasonNumber: Long
+    val tvShowId: Int,
+    val seasonId: Int,
+    val seasonNumber: Int
 )
